Boolean should become Int64

+++ polars/polars-lazy/polars-plan/src/dsl/mod.rs
@@ -900,7 +900,7 @@ impl Expr {
             GetOutput::map_dtype(|dt| {
                 use DataType::*;
                 match dt {
-                    Boolean => UInt32,
+                    Boolean => Int64,
                     UInt64 => UInt64,
                     Float32 => Float32,
                     Float64 => Float64,
